requirements.txt
torch
transformers
datasets
peft
trl
bitsandbytes
biopython
accelerate
requests
jupyterlab
ipywidgets
